\name{testpackage-package}
\alias{testpackage-package}
\alias{testpackage}
\docType{package}
\title{
\packageTitle{testpackage}
}
\description{
\packageDescription{testpackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{testpackage}
\packageIndices{testpackage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{testpackage}

Maintainer: \packageMaintainer{testpackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
